feat: Make reversal writing more DRY, and gate signal handler behind feature flag
