+++ enterprise_subsidy/apps/api_client/tests/test_enterprise.py
@@ -222,7 +222,7 @@ class EnterpriseApiClientTests(TestCase):
                     'enterprise_customer_user': 10,
                     'course_id': self.courserun_key,
                     'created': '2023-05-25T19:27:29Z',
-                    'unenrolled_at': '2023-06-1T19:27:29Z',
+                    'unenrolled_at': '2023-06-01T19:27:29Z',
                 },
                 'transaction_id': transaction_id,
                 'uuid': fulfillment_uuid,
@@ -236,7 +236,7 @@ class EnterpriseApiClientTests(TestCase):
                 'enterprise_customer_user': 10,
                 'course_id': self.courserun_key,
                 'created': '2023-05-25T19:27:29Z',
-                'unenrolled_at': '2023-06-1T19:27:29Z',
+                'unenrolled_at': '2023-06-01T19:27:29Z',
             },
             'transaction_id': transaction_id,
             'uuid': fulfillment_uuid,
